requirements_minimal.txt
cinemagoer==2023.5.1
flask==2.3.3
fuzzywuzzy==0.18.0
google-api-core==2.11.1
google-auth==2.23.0
google-cloud-core==2.3.3
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
gunicorn==21.2.0
mutagen==1.47.0
numpy==1.24.4
pandas==2.0.3
redis==5.0.0
requests==2.31.0
typing-extensions==4.7.1
urllib3==1.26.16
werkzeug==2.3.7
